requirements.txt
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.31.0
openai==1.51.0
tiktoken==0.7.0
pinecone==5.0.0
langchain==0.3.1
langchain-pinecone==0.2.0
langchain-openai==0.2.1
langchain-community==0.3.1
